Update examples, now they are using MasqCrypto.AES vs AES
