+++ index.html
@@ -24,7 +24,7 @@ tr
 <body >
   <h1>Crypto</h1>
   <h2> <a href="./test/index.html">Go to the test page</a></h2>
-  <h2> <a href="./src/example/pfs.html">Go to the PFS demo</a></h2>
+  <h2> <a href="./src/example/pfs/pfs.html">Go to the PFS demo</a></h2>
   <h2> <a href="./src/example/perf/perf.html">Go to the ECDSA and RSA-PSS performance comparison</a></h2>
