+++ src/example/perf/perf.html
@@ -33,6 +33,8 @@ tr
         <th> <button type="button" id="ecPerf">Run Test</button></th>
     </tr>
     </table>
+    
     <script type="text/javascript" src="../../../dist/MasqCrypto.js"></script>
+    <script type="text/javascript" src="../../../dist/example/perf/index.js"></script>
 </body>
 </html>
