+++ src/example/pfs/pfs.html
@@ -81,6 +81,8 @@ tr
           </th>
       </tr>
     </table>
-    <script type="text/javascript" src="../../dist/MasqCrypto.js"></script>
+    <script type="text/javascript" src="localforage.min.js"></script>
+    <script type="text/javascript" src="../../../dist/MasqCrypto.js"></script>
+    <script type="text/javascript" src="../../../dist/example/pfs/index.js"></script>
 </body>
 </html>
